added image controller

+++ market/src/main/java/com/market/domain/ImageRepository.java
@@ -2,5 +2,5 @@ package com.market.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ImageRepository extends CrudRepository<Image, Float> {
+public interface ImageRepository extends CrudRepository<Image, Long> {
 }
